refactor: consolidate duplicated logic in project and notification services
