+++ backend/internal/services/vulnerability_service.go
@@ -4,7 +4,6 @@ import (
 	"archive/tar"
 	"bytes"
 	"context"
-	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -12,10 +11,8 @@ import (
 	"log/slog"
 	"math"
 	"net"
-	"net/url"
 	"os"
 	"path/filepath"
-	"runtime"
 	"sort"
 	"strconv"
 	"strings"
@@ -29,6 +26,7 @@ import (
 	"github.com/getarcaneapp/arcane/backend/v2/pkg/libarcane"
 	activitylib "github.com/getarcaneapp/arcane/backend/v2/pkg/libarcane/activity"
 	"github.com/getarcaneapp/arcane/backend/v2/pkg/libarcane/timeouts"
+	vuln "github.com/getarcaneapp/arcane/backend/v2/pkg/libarcane/vuln"
 	"github.com/getarcaneapp/arcane/backend/v2/pkg/pagination"
 	"github.com/getarcaneapp/arcane/backend/v2/pkg/utils"
 	"github.com/getarcaneapp/arcane/types/v2/vulnerability"
@@ -36,7 +34,6 @@ import (
 	containertypes "github.com/moby/moby/api/types/container"
 	imagetypes "github.com/moby/moby/api/types/image"
 	mounttypes "github.com/moby/moby/api/types/mount"
-	dockerregistry "github.com/moby/moby/api/types/registry"
 	"github.com/moby/moby/client"
 	libupdater "go.getarcane.app/updater/pkg/labels"
 	"gorm.io/gorm"
@@ -49,28 +46,24 @@ const (
 	DefaultArcaneToolsImage = "ghcr.io/getarcaneapp/tools:latest"
 	// DefaultTrivyImage preserves the existing setting name, but now points at
 	// the shared Arcane tools image that includes the Trivy binary.
-	DefaultTrivyImage                        = DefaultArcaneToolsImage
-	DefaultTrivyNetworkMode                  = "bridge"
-	trivyCacheVolumeName                     = "arcane-trivy-cache"
-	trivyCacheMountTarget                    = "/root/.cache"
-	trivyCacheSlotRootDir                    = trivyCacheMountTarget + "/arcane-scan-slots"
-	scanStaleTimeout                         = 30 * time.Minute
-	defaultTrivyCPULimitCores                = 1.0
-	defaultTrivyMemoryLimitMB          int64 = 0
-	trivyNanoCPUsPerCore                     = int64(1_000_000_000)
-	trivyBytesPerMB                          = int64(1024 * 1024)
-	trivyErrorExcerptSize                    = int64(32 * 1024)
-	defaultScanSaveRetryAttempts             = 3
-	defaultScanSaveRetryDelay                = 200 * time.Millisecond
-	defaultScanSaveRetryMaxDelay             = 2 * time.Second
-	trivyOutputPathPrefix                    = "/tmp/arcane-trivy-result-"
-	defaultDockerHostURI                     = "unix:///var/run/docker.sock"
-	DefaultTrivyDBRepository                 = "ghcr.io/getarcaneapp/trivy-db:2"
-	DefaultTrivyJavaDBRepository             = "ghcr.io/getarcaneapp/trivy-java-db:1"
-	DefaultTrivyChecksBundleRepository       = "ghcr.io/getarcaneapp/trivy-checks:1"
-	trivyRegistryConfigDir                   = "/tmp/arcane-registry-auth"
-	trivyRegistryConfigCopyDest              = "/tmp"
-	trivyRegistryConfigTarName               = "arcane-registry-auth/config.json"
+	DefaultTrivyImage                  = DefaultArcaneToolsImage
+	DefaultTrivyNetworkMode            = "bridge"
+	trivyCacheVolumeName               = "arcane-trivy-cache"
+	trivyCacheMountTarget              = "/root/.cache"
+	trivyCacheSlotRootDir              = trivyCacheMountTarget + "/arcane-scan-slots"
+	scanStaleTimeout                   = 30 * time.Minute
+	defaultTrivyCPULimitCores          = 1.0
+	defaultTrivyMemoryLimitMB    int64 = 0
+	trivyNanoCPUsPerCore               = int64(1_000_000_000)
+	trivyBytesPerMB                    = int64(1024 * 1024)
+	trivyErrorExcerptSize              = int64(32 * 1024)
+	defaultScanSaveRetryAttempts       = 3
+	defaultScanSaveRetryDelay          = 200 * time.Millisecond
+	defaultScanSaveRetryMaxDelay       = 2 * time.Second
+	trivyOutputPathPrefix              = "/tmp/arcane-trivy-result-"
+	trivyRegistryConfigDir             = "/tmp/arcane-registry-auth"
+	trivyRegistryConfigCopyDest        = "/tmp"
+	trivyRegistryConfigTarName         = "arcane-registry-auth/config.json"
 )
 
 type trivyRuntimeOptions struct {
@@ -255,36 +248,6 @@ func NewVulnerabilityService(db *database.DB, dockerService *DockerClientService
 	}
 }
 
-func buildDockerConfigJSONInternal(authConfigs map[string]dockerregistry.AuthConfig) ([]byte, error) {
-	if len(authConfigs) == 0 {
-		return nil, nil
-	}
-
-	type authEntry struct {
-		Auth string `json:"auth"`
-	}
-
-	auths := make(map[string]authEntry, len(authConfigs))
-	for host, cfg := range authConfigs {
-		host = strings.TrimSpace(host)
-		if host == "" || cfg.Username == "" || cfg.Password == "" {
-			continue
-		}
-
-		auths[host] = authEntry{
-			Auth: base64.StdEncoding.EncodeToString([]byte(cfg.Username + ":" + cfg.Password)),
-		}
-	}
-
-	if len(auths) == 0 {
-		return nil, nil
-	}
-
-	return json.Marshal(struct {
-		Auths map[string]authEntry `json:"auths"`
-	}{Auths: auths})
-}
-
 func (s *VulnerabilityService) buildTrivyRegistryConfigInternal(ctx context.Context) []byte {
 	if s.registryService == nil {
 		return nil
@@ -296,7 +259,7 @@ func (s *VulnerabilityService) buildTrivyRegistryConfigInternal(ctx context.Cont
 		return nil
 	}
 
-	configJSON, err := buildDockerConfigJSONInternal(authConfigs)
+	configJSON, err := vuln.BuildDockerConfigJSON(authConfigs)
 	if err != nil {
 		slog.WarnContext(ctx, "failed to build trivy registry config; continuing anonymously", "error", err)
 		return nil
@@ -1555,8 +1518,8 @@ func (s *VulnerabilityService) execTrivyScanInContainer(ctx context.Context, con
 
 	execCmd := make([]string, 0, 16)
 	execCmd = append(execCmd, "trivy", "image", "--format", "json", "--quiet", "--output", outputPath, "--timeout", trivyTimeout)
-	execCmd = append(execCmd, trivyScanCacheBackendArgsInternal()...)
-	execCmd = append(execCmd, trivyDefaultRepositoryArgsInternal()...)
+	execCmd = append(execCmd, vuln.ScanCacheBackendArgs()...)
+	execCmd = append(execCmd, vuln.DefaultRepositoryArgs()...)
 	execCmd = append(execCmd, imageName)
 
 	execCfg := client.ExecCreateOptions{
@@ -1801,57 +1764,11 @@ func (s *VulnerabilityService) GetTrivyVersion(ctx context.Context) string {
 		return ""
 	}
 
-	return parseTrivyVersion(readTempFileExcerptInternal(stdoutFile))
-}
-
-func parseTrivyVersion(output string) string {
-	output = strings.TrimSpace(output)
-	if output == "" {
-		return ""
-	}
-	lines := strings.SplitSeq(output, "\n")
-	for line := range lines {
-		if after, ok := strings.CutPrefix(line, "Version:"); ok {
-			return strings.TrimSpace(after)
-		}
-	}
-	return strings.TrimSpace(output)
+	return vuln.ParseVersion(readTempFileExcerptInternal(stdoutFile))
 }
 
 func (s *VulnerabilityService) getTrivyImageRef() string {
 	return DefaultTrivyImage
-}
-
-func normalizeTrivyNetworkModeInternal(networkMode string) string {
-	return strings.TrimSpace(networkMode)
-}
-
-func parseTrivySecurityOptsInternal(value string) []string {
-	value = strings.TrimSpace(value)
-	if value == "" {
-		return nil
-	}
-
-	value = strings.NewReplacer("\r\n", "\n", "\r", "\n").Replace(value)
-	parts := strings.FieldsFunc(value, func(r rune) bool {
-		return r == ',' || r == '\n'
-	})
-	if len(parts) == 0 {
-		return nil
-	}
-
-	opts := make([]string, 0, len(parts))
-	for _, part := range parts {
-		if opt := strings.TrimSpace(part); opt != "" {
-			opts = append(opts, opt)
-		}
-	}
-
-	if len(opts) == 0 {
-		return nil
-	}
-
-	return opts
 }
 
 func (s *VulnerabilityService) getConfiguredTrivyNetworkModeInternal() string {
@@ -1864,7 +1781,7 @@ func (s *VulnerabilityService) getConfiguredTrivyNetworkModeInternal() string {
 		return ""
 	}
 
-	return normalizeTrivyNetworkModeInternal(cfg.TrivyNetwork.Value)
+	return vuln.NormalizeNetworkMode(cfg.TrivyNetwork.Value)
 }
 
 func (s *VulnerabilityService) getEffectiveDockerHostInternal() string {
@@ -1882,46 +1799,7 @@ func (s *VulnerabilityService) getEffectiveDockerHostInternal() string {
 		}
 	}
 
-	return defaultDockerHostURI
-}
-
-func parseTrivyDockerHostInternal(dockerHost string) (scheme string, socketPath string, err error) {
-	dockerHost = strings.TrimSpace(dockerHost)
-	if dockerHost == "" {
-		dockerHost = defaultDockerHostURI
-	}
-
-	if strings.HasPrefix(dockerHost, "/") {
-		return "unix", dockerHost, nil
-	}
-
-	parsed, err := url.Parse(dockerHost)
-	if err != nil {
-		return "", "", fmt.Errorf("parse docker host %q: %w", dockerHost, err)
-	}
-
-	scheme = strings.ToLower(strings.TrimSpace(parsed.Scheme))
-	switch scheme {
-	case "unix":
-		socketPath = strings.TrimSpace(parsed.Path)
-		if socketPath == "" {
-			return "", "", fmt.Errorf("docker host %q is missing a unix socket path", dockerHost)
-		}
-		return scheme, socketPath, nil
-	case "tcp", "http", "https":
-		return scheme, "", nil
-	default:
-		return "", "", fmt.Errorf("unsupported docker host scheme %q", scheme)
-	}
-}
-
-func buildTrivyDockerHostEnvInternal(dockerHost string) []string {
-	dockerHost = strings.TrimSpace(dockerHost)
-	if dockerHost == "" {
-		return nil
-	}
-
-	return []string{"DOCKER_HOST=" + dockerHost}
+	return vuln.DefaultDockerHostURI
 }
 
 func resolveTrivyUnixSocketSourceInternal(
@@ -2029,7 +1907,7 @@ func selectTrivyAutoNetworkModeInternal(inspect *containertypes.InspectResponse)
 	}
 
 	if inspect.HostConfig != nil {
-		networkMode := normalizeTrivyNetworkModeInternal(string(inspect.HostConfig.NetworkMode))
+		networkMode := vuln.NormalizeNetworkMode(string(inspect.HostConfig.NetworkMode))
 		if networkMode != "" && networkMode != "default" && networkMode != DefaultTrivyNetworkMode &&
 			!containertypes.NetworkMode(networkMode).IsContainer() {
 			return networkMode
@@ -2039,7 +1917,7 @@ func selectTrivyAutoNetworkModeInternal(inspect *containertypes.InspectResponse)
 	if inspect.NetworkSettings != nil && len(inspect.NetworkSettings.Networks) > 0 {
 		networkNames := make([]string, 0, len(inspect.NetworkSettings.Networks))
 		for networkName := range inspect.NetworkSettings.Networks {
-			networkName = normalizeTrivyNetworkModeInternal(networkName)
+			networkName = vuln.NormalizeNetworkMode(networkName)
 			if networkName != "" {
 				networkNames = append(networkNames, networkName)
 			}
@@ -2063,7 +1941,7 @@ func selectTrivyAutoNetworkModeInternal(inspect *containertypes.InspectResponse)
 	}
 
 	if inspect.HostConfig != nil {
-		networkMode := normalizeTrivyNetworkModeInternal(string(inspect.HostConfig.NetworkMode))
+		networkMode := vuln.NormalizeNetworkMode(string(inspect.HostConfig.NetworkMode))
 		if networkMode != "" && networkMode != "default" && !containertypes.NetworkMode(networkMode).IsContainer() {
 			return networkMode
 		}
@@ -2092,12 +1970,12 @@ func (s *VulnerabilityService) resolveTrivyRuntimeOptionsInternal(ctx context.Co
 	}
 
 	dockerHost := s.getEffectiveDockerHostInternal()
-	scheme, socketPath, err := parseTrivyDockerHostInternal(dockerHost)
+	scheme, socketPath, err := vuln.ParseDockerHost(dockerHost)
 	if err != nil {
 		return trivyRuntimeOptions{}, fmt.Errorf("failed to resolve trivy docker host: %w", err)
 	}
 
-	options.ContainerEnv = buildTrivyDockerHostEnvInternal(dockerHost)
+	options.ContainerEnv = vuln.BuildDockerHostEnv(dockerHost)
 	if scheme != "unix" {
 		return options, nil
 	}
@@ -2136,7 +2014,7 @@ func (s *VulnerabilityService) getTrivyRuntimeSecurityInternal() ([]string, bool
 		return nil, false
 	}
 
-	return parseTrivySecurityOptsInternal(cfg.TrivySecurityOpts.Value), cfg.TrivyPrivileged.IsTrue()
+	return vuln.ParseSecurityOpts(cfg.TrivySecurityOpts.Value), cfg.TrivyPrivileged.IsTrue()
 }
 
 func (s *VulnerabilityService) getTrivyScanTimeoutArg() string {
@@ -2277,29 +2155,6 @@ func cleanupTempFiles(ctx context.Context, tempFiles []string) {
 
 // trivyScanCacheBackendArgsForArchInternal returns extra Trivy flags that switch
 // the scan cache to Trivy's documented in-memory backend on 32-bit architectures.
-// The default BoltDB-backed cache can fail with ENOMEM on arm/v7 and 386 because
-// Go's heap reservations fragment the limited 32-bit virtual address space.
-func trivyScanCacheBackendArgsForArchInternal(arch string) []string {
-	switch arch {
-	case "arm", "386", "mips", "mipsle":
-		return []string{"--cache-backend", "memory"}
-	default:
-		return nil
-	}
-}
-
-func trivyScanCacheBackendArgsInternal() []string {
-	return trivyScanCacheBackendArgsForArchInternal(runtime.GOARCH)
-}
-
-func trivyDefaultRepositoryArgsInternal() []string {
-	return []string{
-		"--db-repository", DefaultTrivyDBRepository,
-		"--java-db-repository", DefaultTrivyJavaDBRepository,
-		"--checks-bundle-repository", DefaultTrivyChecksBundleRepository,
-	}
-}
-
 func (s *VulnerabilityService) buildTrivyCommandArgs(
 	ctx context.Context,
 	imageName string,
@@ -2309,8 +2164,8 @@ func (s *VulnerabilityService) buildTrivyCommandArgs(
 	outputPath string,
 ) ([]string, []string) {
 	cmdArgs := []string{"image", "--format", "json", "--quiet", "--timeout", s.getTrivyScanTimeoutArg()}
-	cmdArgs = append(cmdArgs, trivyScanCacheBackendArgsInternal()...)
-	cmdArgs = append(cmdArgs, trivyDefaultRepositoryArgsInternal()...)
+	cmdArgs = append(cmdArgs, vuln.ScanCacheBackendArgs()...)
+	cmdArgs = append(cmdArgs, vuln.DefaultRepositoryArgs()...)
 	var tempFiles []string
 
 	if strings.TrimSpace(configContent) != "" {
@@ -2338,18 +2193,6 @@ func (s *VulnerabilityService) buildTrivyCommandArgs(
 	cmdArgs = append(cmdArgs, imageName)
 
 	return cmdArgs, tempFiles
-}
-
-func buildTrivyContainerConfig(trivyImage string, cmdArgs []string, env []string) *containertypes.Config {
-	return &containertypes.Config{
-		Image:      trivyImage,
-		Entrypoint: []string{"trivy"},
-		Cmd:        cmdArgs,
-		Env:        append([]string(nil), env...),
-		Labels: map[string]string{
-			libarcane.InternalResourceLabel: "true",
-		},
-	}
 }
 
 func buildTrivyBatchHostConfigInternal(
@@ -2407,7 +2250,7 @@ func buildTrivyHostConfig(
 }
 
 func selectDefaultTrivyNetworkModeInternal(networkMode string) string {
-	networkMode = normalizeTrivyNetworkModeInternal(networkMode)
+	networkMode = vuln.NormalizeNetworkMode(networkMode)
 	if networkMode == "" {
 		return DefaultTrivyNetworkMode
 	}
@@ -3447,7 +3290,7 @@ func (s *VulnerabilityService) runTrivyScan(ctx context.Context, trivyImage stri
 		env = append(env, "DOCKER_CONFIG="+trivyRegistryConfigDir)
 	}
 
-	config := buildTrivyContainerConfig(trivyImage, cmdArgs, env)
+	config := vuln.BuildContainerConfig(trivyImage, cmdArgs, env)
 	resources, cpuSet, applyLimits := s.getTrivyContainerResourceOptionsInternal(ctx, dockerClient)
 	securityOpts, privileged := s.getTrivyRuntimeSecurityInternal()
 	hostConfig := buildTrivyHostConfig(
