+++ backend/internal/services/vulnerability_service_test.go
@@ -2,7 +2,6 @@ package services
 
 import (
 	"context"
-	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -16,86 +15,16 @@ import (
 	"github.com/getarcaneapp/arcane/backend/v2/internal/database"
 	"github.com/getarcaneapp/arcane/backend/v2/internal/models"
 	"github.com/getarcaneapp/arcane/backend/v2/pkg/libarcane/timeouts"
+	vuln "github.com/getarcaneapp/arcane/backend/v2/pkg/libarcane/vuln"
 	"github.com/getarcaneapp/arcane/backend/v2/pkg/pagination"
 	"github.com/getarcaneapp/arcane/types/v2/vulnerability"
 	glsqlite "github.com/glebarez/sqlite"
 	containertypes "github.com/moby/moby/api/types/container"
 	mounttypes "github.com/moby/moby/api/types/mount"
 	networktypes "github.com/moby/moby/api/types/network"
-	dockerregistry "github.com/moby/moby/api/types/registry"
 	"github.com/stretchr/testify/require"
 	"gorm.io/gorm"
 )
-
-func TestBuildDockerConfigJSONInternal(t *testing.T) {
-	tests := []struct {
-		name      string
-		auths     map[string]dockerregistry.AuthConfig
-		wantAuths map[string]string
-	}{
-		{
-			name: "nil map",
-		},
-		{
-			name:  "empty map",
-			auths: map[string]dockerregistry.AuthConfig{},
-		},
-		{
-			name: "one registry",
-			auths: map[string]dockerregistry.AuthConfig{
-				"registry.example.com": {Username: "user", Password: "pass"},
-			},
-			wantAuths: map[string]string{
-				"registry.example.com": base64.StdEncoding.EncodeToString([]byte("user:pass")),
-			},
-		},
-		{
-			name: "skips blank credentials",
-			auths: map[string]dockerregistry.AuthConfig{
-				"registry.example.com": {Username: "user", Password: "pass"},
-				"blank-user.example":   {Username: "", Password: "pass"},
-				"blank-pass.example":   {Username: "user", Password: ""},
-				"   ":                  {Username: "user", Password: "pass"},
-			},
-			wantAuths: map[string]string{
-				"registry.example.com": base64.StdEncoding.EncodeToString([]byte("user:pass")),
-			},
-		},
-		{
-			name: "two registries",
-			auths: map[string]dockerregistry.AuthConfig{
-				"registry-a.example.com": {Username: "user-a", Password: "pass-a"},
-				"registry-b.example.com": {Username: "user-b", Password: "pass-b"},
-			},
-			wantAuths: map[string]string{
-				"registry-a.example.com": base64.StdEncoding.EncodeToString([]byte("user-a:pass-a")),
-				"registry-b.example.com": base64.StdEncoding.EncodeToString([]byte("user-b:pass-b")),
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := buildDockerConfigJSONInternal(tt.auths)
-			require.NoError(t, err)
-			if len(tt.wantAuths) == 0 {
-				require.Nil(t, got)
-				return
-			}
-
-			var parsed struct {
-				Auths map[string]struct {
-					Auth string `json:"auth"`
-				} `json:"auths"`
-			}
-			require.NoError(t, json.Unmarshal(got, &parsed))
-			require.Len(t, parsed.Auths, len(tt.wantAuths))
-			for host, want := range tt.wantAuths {
-				require.Equal(t, want, parsed.Auths[host].Auth)
-			}
-		})
-	}
-}
 
 func TestIsExpectedDockerStreamEndErrorInternal(t *testing.T) {
 	tests := []struct {
@@ -367,49 +296,11 @@ func TestBuildTrivyCommandArgs_IncludesDefaultRepositories(t *testing.T) {
 	cmdArgs, _ := svc.buildTrivyCommandArgs(context.Background(), "alpine:3.20", "", "", "", "/tmp/out.json")
 
 	require.Contains(t, cmdArgs, "--db-repository")
-	require.Contains(t, cmdArgs, DefaultTrivyDBRepository)
+	require.Contains(t, cmdArgs, vuln.DefaultDBRepository)
 	require.Contains(t, cmdArgs, "--java-db-repository")
-	require.Contains(t, cmdArgs, DefaultTrivyJavaDBRepository)
+	require.Contains(t, cmdArgs, vuln.DefaultJavaDBRepository)
 	require.Contains(t, cmdArgs, "--checks-bundle-repository")
-	require.Contains(t, cmdArgs, DefaultTrivyChecksBundleRepository)
-}
-
-func TestTrivyScanCacheBackendArgsForArch(t *testing.T) {
-	memoryBackend := []string{"--cache-backend", "memory"}
-
-	tests := []struct {
-		arch string
-		want []string
-	}{
-		{"arm", memoryBackend},
-		{"386", memoryBackend},
-		{"mips", memoryBackend},
-		{"mipsle", memoryBackend},
-		{"amd64", nil},
-		{"arm64", nil},
-		{"ppc64le", nil},
-		{"s390x", nil},
-		{"", nil},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.arch, func(t *testing.T) {
-			got := trivyScanCacheBackendArgsForArchInternal(tt.arch)
-			require.Equal(t, tt.want, got)
-			require.NotContains(t, got, "--db-backend")
-		})
-	}
-}
-
-func TestBuildTrivyContainerConfig_IncludesEnv(t *testing.T) {
-	config := buildTrivyContainerConfig(
-		DefaultTrivyImage,
-		[]string{"image", "alpine:3.20"},
-		[]string{"DOCKER_HOST=tcp://docker-socket-proxy:2375"},
-	)
-
-	require.Equal(t, []string{"trivy"}, config.Entrypoint)
-	require.Equal(t, []string{"DOCKER_HOST=tcp://docker-socket-proxy:2375"}, config.Env)
+	require.Contains(t, cmdArgs, vuln.DefaultChecksBundleRepository)
 }
 
 func TestBuildTrivyHostConfig_ExcludesResourcesWhenLimitsDisabled(t *testing.T) {
@@ -484,27 +375,6 @@ func TestBuildTrivyHostConfig_UsesCPUSetAndClearsNanoCPUs(t *testing.T) {
 	require.Equal(t, resources.MemorySwap, hostConfig.MemorySwap)
 }
 
-func TestParseTrivySecurityOptsInternal(t *testing.T) {
-	tests := []struct {
-		name  string
-		value string
-		want  []string
-	}{
-		{name: "empty", value: "", want: nil},
-		{name: "whitespace only", value: " \n\t ", want: nil},
-		{name: "single", value: "label=disable", want: []string{"label=disable"}},
-		{name: "comma separated", value: "label=disable,label=type:container_runtime_t", want: []string{"label=disable", "label=type:container_runtime_t"}},
-		{name: "newline separated", value: "label=disable\nlabel=type:container_runtime_t", want: []string{"label=disable", "label=type:container_runtime_t"}},
-		{name: "mixed whitespace", value: " label=disable,\n\n  label=type:container_runtime_t \r\n privileged=true ", want: []string{"label=disable", "label=type:container_runtime_t", "privileged=true"}},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			require.Equal(t, tt.want, parseTrivySecurityOptsInternal(tt.value))
-		})
-	}
-}
-
 func TestBuildTrivyHostConfig_AppliesRuntimeSecurity(t *testing.T) {
 	hostConfig := buildTrivyHostConfig(
 		"cache-volume",
@@ -572,48 +442,10 @@ func TestApplyTrivyRuntimeSecurityInternal(t *testing.T) {
 	require.True(t, hostConfig.Privileged)
 }
 
-func TestNormalizeTrivyNetworkModeInternal(t *testing.T) {
-	require.Equal(t, "", normalizeTrivyNetworkModeInternal(""))
-	require.Equal(t, "", normalizeTrivyNetworkModeInternal(" \t\n "))
-	require.Equal(t, "arcane-external", normalizeTrivyNetworkModeInternal("arcane-external"))
-}
-
 func TestSelectDefaultTrivyNetworkModeInternal(t *testing.T) {
 	require.Equal(t, DefaultTrivyNetworkMode, selectDefaultTrivyNetworkModeInternal(""))
 	require.Equal(t, DefaultTrivyNetworkMode, selectDefaultTrivyNetworkModeInternal(" \t\n "))
 	require.Equal(t, "arcane-external", selectDefaultTrivyNetworkModeInternal("arcane-external"))
-}
-
-func TestParseTrivyDockerHostInternal(t *testing.T) {
-	tests := []struct {
-		name           string
-		dockerHost     string
-		wantScheme     string
-		wantSocketPath string
-		wantErr        string
-	}{
-		{name: "default unix", dockerHost: "", wantScheme: "unix", wantSocketPath: "/var/run/docker.sock"},
-		{name: "explicit unix", dockerHost: "unix:///run/user/1000/podman/podman.sock", wantScheme: "unix", wantSocketPath: "/run/user/1000/podman/podman.sock"},
-		{name: "tcp host", dockerHost: "tcp://docker-socket-proxy:2375", wantScheme: "tcp"},
-		{name: "http host", dockerHost: "http://docker-socket-proxy:2375", wantScheme: "http"},
-		{name: "https host", dockerHost: "https://docker.example.com", wantScheme: "https"},
-		{name: "unsupported scheme", dockerHost: "ssh://docker.example.com", wantErr: "unsupported docker host scheme"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			scheme, socketPath, err := parseTrivyDockerHostInternal(tt.dockerHost)
-			if tt.wantErr != "" {
-				require.Error(t, err)
-				require.Contains(t, err.Error(), tt.wantErr)
-				return
-			}
-
-			require.NoError(t, err)
-			require.Equal(t, tt.wantScheme, scheme)
-			require.Equal(t, tt.wantSocketPath, socketPath)
-		})
-	}
 }
 
 func TestResolveTrivyUnixSocketSourceInternal(t *testing.T) {
